move csr_reader and cif_io
