+++ source/source_base/module_out/csr_reader.h
@@ -3,8 +3,8 @@
 
 #include <fstream>
 
-#include "source_base/module_out/file_reader.h"
-#include "source_base/module_out/sparse_matrix.h"
+#include "file_reader.h"
+#include "sparse_matrix.h"
 
 namespace ModuleIO
 {
