+++ source/source_base/module_out/orb_io.cpp
@@ -1,4 +1,4 @@
-#include "source_io/module_output/orb_io.h"
+#include "orb_io.h"
 #include "source_base/tool_quit.h"
 #ifdef __MPI
 #include "source_base/parallel_common.h"
