+++ source/source_basis/module_nao/test/CMakeLists.txt
@@ -16,7 +16,7 @@ AddTest(
     ../numerical_radial.cpp
     ../../module_ao/ORB_atomic_lm.cpp
     ../../module_ao/ORB_atomic.cpp
-    ../../../source_io/module_output/orb_io.cpp
+    
   LIBS parameter device base
 )
 
@@ -29,7 +29,7 @@ AddTest(
     ../numerical_radial.cpp
     ../../module_ao/ORB_atomic_lm.cpp
     ../../module_ao/ORB_atomic.cpp
-    ../../../source_io/module_output/orb_io.cpp
+    
   LIBS parameter device base
 )
 
@@ -42,7 +42,7 @@ AddTest(
     ../numerical_radial.cpp
     ../../module_ao/ORB_atomic_lm.cpp
     ../../module_ao/ORB_atomic.cpp
-    ../../../source_io/module_output/orb_io.cpp
+    
   LIBS parameter device base
 )
 
@@ -55,7 +55,7 @@ AddTest(
     ../numerical_radial.cpp
     ../../module_ao/ORB_atomic_lm.cpp
     ../../module_ao/ORB_atomic.cpp
-    ../../../source_io/module_output/orb_io.cpp
+    
   LIBS parameter device base
 )
 
@@ -68,7 +68,7 @@ AddTest(
     ../numerical_radial.cpp
     ../../module_ao/ORB_atomic_lm.cpp
     ../../module_ao/ORB_atomic.cpp
-    ../../../source_io/module_output/orb_io.cpp
+    
  LIBS parameter device base
 )
 
@@ -86,7 +86,7 @@ AddTest(
     ../sphbes_radials.cpp
     ../../module_ao/ORB_atomic_lm.cpp
     ../../module_ao/ORB_atomic.cpp
-    ../../../source_io/module_output/orb_io.cpp
+    
   LIBS parameter device base
 )
 
@@ -106,7 +106,7 @@ AddTest(
     ../two_center_bundle.cpp
     ../two_center_integrator.cpp
     ../real_gaunt_table.cpp
-    ../../../source_io/module_output/orb_io.cpp
+    
   LIBS parameter device base container orb
 )
 
@@ -135,7 +135,7 @@ AddTest(
     ../radial_set.cpp
     ../numerical_radial.cpp
     ../two_center_bundle.cpp
-    ../../../source_io/module_output/orb_io.cpp
+    
   LIBS parameter device base container orb
 )
 
@@ -155,7 +155,7 @@ AddTest(
     ../sphbes_radials.cpp
     ../radial_set.cpp
     ../numerical_radial.cpp
-    ../../../source_io/module_output/orb_io.cpp
+    
   LIBS parameter device base container orb
 )
 
@@ -175,7 +175,7 @@ AddTest(
     ../sphbes_radials.cpp
     ../radial_set.cpp
     ../numerical_radial.cpp
-    ../../../source_io/module_output/orb_io.cpp
+    
   LIBS parameter device base container orb
 )
