+++ source/source_esolver/test/CMakeLists.txt
@@ -23,6 +23,6 @@ AddTest(
     esolver_dp_test.cpp
     ../esolver_dp.cpp
     ../../source_cell/base_cell.cpp
-    ../../source_io/module_output/cif_io.cpp
+    ../../source_cell/cif_io.cpp
     ../../source_io/module_output/output_log.cpp
 )
