+++ source/source_io/test/cif_io_test.cpp
@@ -1,5 +1,5 @@
 #include <gtest/gtest.h>
-#include "source_io/module_output/cif_io.h"
+#include "source_cell/cif_io.h"
 #include <cmath>
 #include <random>
 #include "source_base/formatter.h"
