+++ source/source_io/test/orb_io_test.cpp
@@ -1,5 +1,5 @@
 #include <gtest/gtest.h>
-#include "source_io/module_output/orb_io.h"
+#include "source_base/module_out/orb_io.h"
 
 #ifdef __MPI
 #include <mpi.h>
