+++ source/source_io/test/write_hs_r_compat_test.cpp
@@ -12,7 +12,7 @@
 #include "source_io/module_hs/rr_sparse_writer.h"
 #include "source_io/module_hs/write_HS_R.h"
 #include "source_io/module_hs/write_HS_sparse.h"
-#include "source_io/module_output/csr_reader.h"
+#include "source_base/module_out/csr_reader.h"
 #include "source_hamilt/module_hcontainer/atom_pair.h"
 #include "source_hamilt/module_hcontainer/hcontainer.h"
