+++ source/source_lcao/module_lr/utils/test/CMakeLists.txt
@@ -3,7 +3,7 @@ AddTest(
   TARGET MODULE_LR_lr_util_phys_test
   LIBS parameter base device container planewave #for FFT
   SOURCES lr_util_physics_test.cpp  ../lr_util.cpp
-  ../../../../source_io/module_output/orb_io.cpp
+  
 )
 
 AddTest(
