+++ source/source_lcao/module_ri/Exx_LRI_interface.hpp
@@ -6,7 +6,7 @@
 #include "source_estate/elecstate_lcao.h"
 #include "source_hamilt/module_xc/xc_functional.h"
 #include "source_io/module_hs/write_HS_sparse.h"
-#include "source_io/module_output/csr_reader.h"
+#include "source_base/module_out/csr_reader.h"
 #include "source_io/module_parameter/parameter.h"
 #include "source_io/module_restart/restart.h"
 #include "source_io/module_restart/restart_exx_csr.h"
