+++ source/source_relax/relax_driver.cpp
@@ -2,11 +2,11 @@
 #include "source_base/formatter.h"
 #include "source_base/global_file.h"
 #include "source_base/version.h"
-#include "source_io/module_output/cif_io.h"
+#include "source_cell/cif_io.h"
 #include "source_io/module_json/output_info.h"
 #include "source_io/module_output/output_log.h"
 #include "source_io/module_output/print_info.h"
-#include "source_io/module_output/read_exit_file.h"
+#include "source_base/module_out/read_exit_file.h"
 #include "source_io/module_parameter/parameter.h"
 #include "source_cell/print_cell.h"
